requirements.txt
torch>=2.0.0
transformers>=4.36.0
pillow>=12.0.0